+++ migrate/versioning/api.py
@@ -1,12 +1,17 @@
 """
    This module provides an external API to the versioning system.
 
-   .. versionchanged:: 0.4.5 
+   .. versionchanged:: 0.6.0
+   :func:`migrate.versioning.api.test` and schema diff functions \
+   changed order of positional arguments so all accept `url` and `repository`\
+   as first arguments.
+
+   .. versionchanged:: 0.5.4 
     ``--preview_sql`` displays source file when using SQL scripts.
     If Python script is used, it runs the action with mocked engine and
     returns captured SQL statements.
 
-   .. versionchanged:: 0.4.5
+   .. versionchanged:: 0.5.4
     Deprecated ``--echo`` parameter in favour of new
     :func:`migrate.versioning.util.construct_engine` behavior.
 """
@@ -74,6 +79,7 @@ def help(cmd=None, **opts):
         ret = ret.replace('%prog', sys.argv[0])
     return ret
 
+
 @catch_known_errors
 def create(repository, name, **opts):
     """%prog create REPOSITORY_PATH NAME [--table=TABLE]
@@ -84,7 +90,7 @@ def create(repository, name, **opts):
     'migrate_version'.  This table is created in all version-controlled
     databases.
     """
-    repo_path = Repository.create(repository, name, **opts)
+    Repository.create(repository, name, **opts)
 
 
 @catch_known_errors
@@ -192,8 +198,8 @@ def downgrade(url, repository, version, **opts):
         "Try 'upgrade' instead."
     return _migrate(url, repository, version, upgrade=False, err=err, **opts)
     
-def test(repository, url, **opts):
-    """%prog test REPOSITORY_PATH URL [VERSION]
+def test(url, repository, **opts):
+    """%prog test URL REPOSITORY_PATH [VERSION]
 
     Performs the upgrade and downgrade option on the given
     database. This is not a real test and may leave the database in a
@@ -267,8 +273,8 @@ def manage(file, **opts):
     return Repository.create_manage_file(file, **opts)
 
 
-def compare_model_to_db(url, model, repository, **opts):
-    """%prog compare_model_to_db URL MODEL REPOSITORY_PATH
+def compare_model_to_db(url, repository, model, **opts):
+    """%prog compare_model_to_db URL REPOSITORY_PATH MODEL
 
     Compare the current model (assumed to be a module level variable
     of type sqlalchemy.MetaData) against the current database.
@@ -276,7 +282,7 @@ def compare_model_to_db(url, model, repository, **opts):
     NOTE: This is EXPERIMENTAL.
     """  # TODO: get rid of EXPERIMENTAL label
     engine = construct_engine(url, **opts)
-    print ControlledSchema.compare_model_to_db(engine, model, repository)
+    return ControlledSchema.compare_model_to_db(engine, model, repository)
 
 
 def create_model(url, repository, **opts):
@@ -288,13 +294,12 @@ def create_model(url, repository, **opts):
     """  # TODO: get rid of EXPERIMENTAL label
     engine = construct_engine(url, **opts)
     declarative = opts.get('declarative', False)
-    print ControlledSchema.create_model(engine, repository, declarative)
+    return ControlledSchema.create_model(engine, repository, declarative)
 
 
-# TODO: get rid of this? if we don't add back path param
 @catch_known_errors
-def make_update_script_for_model(url, oldmodel, model, repository, **opts):
-    """%prog make_update_script_for_model URL OLDMODEL MODEL REPOSITORY_PATH
+def make_update_script_for_model(url, repository, oldmodel, model, **opts):
+    """%prog make_update_script_for_model URL REPOSITORY_PATH OLDMODEL MODEL
 
     Create a script changing the old Python model to the new (current)
     Python model, sending to stdout.
@@ -302,12 +307,12 @@ def make_update_script_for_model(url, oldmodel, model, repository, **opts):
     NOTE: This is EXPERIMENTAL.
     """  # TODO: get rid of EXPERIMENTAL label
     engine = construct_engine(url, **opts)
-    print PythonScript.make_update_script_for_model(
+    return PythonScript.make_update_script_for_model(
         engine, oldmodel, model, repository, **opts)
 
 
-def update_db_from_model(url, model, repository, **opts):
-    """%prog update_db_from_model URL MODEL REPOSITORY_PATH
+def update_db_from_model(url, repository, model, **opts):
+    """%prog update_db_from_model URL REPOSITORY_PATH MODEL
 
     Modify the database to match the structure of the current Python
     model. This also sets the db_version number to the latest in the
@@ -337,15 +342,14 @@ def _migrate(url, repository, version, upgrade, err, **opts):
                 print change.source()
 
         elif opts.get('preview_py'):
+            if not isinstance(change, PythonScript):
+                raise exceptions.UsageError("Python source can be only displayed"
+                    " for python migration files")
             source_ver = max(ver, nextver)
             module = schema.repository.version(source_ver).script().module
             funcname = upgrade and "upgrade" or "downgrade"
             func = getattr(module, funcname)
-            if isinstance(change, PythonScript):
-                print inspect.getsource(func)
-            else:
-                raise UsageError("Python source can be only displayed"
-                    " for python migration files")
+            print inspect.getsource(func)
         else:
             schema.runchange(ver, change, changeset.step)
             print 'done'