+++ migrate/versioning/schema.py
@@ -1,6 +1,8 @@
 """
    Database schema version management.
 """
+import sys
+
 from sqlalchemy import (Table, Column, MetaData, String, Text, Integer,
     create_engine)
 from sqlalchemy.sql import and_
@@ -32,22 +34,17 @@ class ControlledSchema(object):
     def load(self):
         """Load controlled schema version info from DB"""
         tname = self.repository.version_table
-        if not hasattr(self, 'table') or self.table is None:
-            try:
-                self.table = Table(tname, self.meta, autoload=True)
-            except (sa_exceptions.NoSuchTableError,
-                    AssertionError):
-                # assertionerror is raised if no table is found in oracle db
-                raise exceptions.DatabaseNotControlledError(tname)
-
-        # TODO?: verify that the table is correct (# cols, etc.)
-        result = self.engine.execute(self.table.select(
-                    self.table.c.repository_id == str(self.repository.id)))
-
         try:
+            if not hasattr(self, 'table') or self.table is None:
+                    self.table = Table(tname, self.meta, autoload=True)
+
+            result = self.engine.execute(self.table.select(
+                self.table.c.repository_id == str(self.repository.id)))
+
             data = list(result)[0]
-        except IndexError:
-            raise exceptions.DatabaseNotControlledError(tname)
+        except Exception:
+            cls, exc, tb = sys.exc_info()
+            raise exceptions.DatabaseNotControlledError, exc.message, tb
 
         self.version = data['version']
         return data