+++ migrate/versioning/script/sql.py
@@ -7,8 +7,16 @@ from migrate.versioning.script import base
 class SqlScript(base.BaseScript):
     """A file containing plain SQL statements."""
 
+    @classmethod
+    def create(cls, path, **opts):
+        """Create an empty migration script at specified path
+        
+        :returns: :class:`SqlScript instance <migrate.versioning.script.sql.SqlScript>`"""
+        cls.require_notfound(path)
+        open(path, "w").close()
+
     # TODO: why is step parameter even here?
-    def run(self, engine, step=None):
+    def run(self, engine, step=None, executemany=True):
         """Runs SQL script through raw dbapi execute call"""
         text = self.source()
         # Don't rely on SA's autocommit here
@@ -21,7 +29,7 @@ class SqlScript(base.BaseScript):
                 # HACK: SQLite doesn't allow multiple statements through
                 # its execute() method, but it provides executescript() instead
                 dbapi = conn.engine.raw_connection()
-                if getattr(dbapi, 'executescript', None):
+                if executemany and getattr(dbapi, 'executescript', None):
                     dbapi.executescript(text)
                 else:
                     conn.execute(text)